+++ src/MagicOnion.Client.SourceGenerator/CodeGen/StaticStreamingHubClientGenerator.cs
@@ -118,7 +118,7 @@ public class StaticStreamingHubClientGenerator
                             public {{ctx.Hub.GetClientFullName()}}(global::Grpc.Core.CallInvoker callInvoker, global::System.String host, global::Grpc.Core.CallOptions options, global::MagicOnion.Serialization.IMagicOnionSerializerProvider serializerProvider, global::MagicOnion.Client.IMagicOnionClientLogger logger)
                                 : base(callInvoker, host, options, serializerProvider, logger)
                             {
-                                var marshaller = global::MagicOnion.MagicOnionMarshallers.ThroughMarshaller;
+                                var marshaller = global::MagicOnion.Internal.MagicOnionMarshallers.ThroughMarshaller;
                                 DuplexStreamingAsyncMethod = new global::Grpc.Core.Method<global::System.Byte[], global::System.Byte[]>(global::Grpc.Core.MethodType.DuplexStreaming, "{{ctx.Hub.ServiceType.Name}}", "Connect", marshaller, marshaller);
                             }
             """);
